% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_datefields.R
\name{add_datefields}
\alias{add_datefields}
\title{Adds ISO or CDC datefields to a data set with a "Date" (Date) field}
\usage{
add_datefields(df, type = "ISO")
}
\value{
dataframe (or tibble) with week and year fields added
}
\description{
Adds ISO or CDC datefields to a data set with a "Date" (Date) field
}
