% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatters_calculators.R
\name{calc_env_anomalies}
\alias{calc_env_anomalies}
\title{Calculates anomalies}
\usage{
calc_env_anomalies(env_ts, quo_groupfield, quo_obsfield, report_dates)
}
\description{
Calculates anomalies
}
