% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatters_calculators.R
\name{calc_return_value}
\alias{calc_return_value}
\title{Calculate the epidemiological value to be shown as result, and on the reports.}
\usage{
calc_return_value(cases, c_quo_tf = FALSE, q_pop = NULL, inc_per, vt,
  mc)
}
\arguments{
\item{cases}{Field containing case counts, if a quosure, c_quo_tf should be TRUE.}

\item{c_quo_tf}{Binary T/F if case field is a quosure rather than a column name.}

\item{q_pop}{Quosure of user-given field containing population values.}

\item{inc_per}{Number for what unit of population the incidence should be
reported in, e.g. incidence rate of 3 per 1000 people. Parameter ignored if
vt == "cases" ("incidence" is default, if not set).}

\item{vt}{From match.arg evaluation of fc_control$value_type, whether to return
epidemiological report values in "incidence" (default) or "cases".}

\item{mc}{From match.arg evaluation of model_choice. Reserved for future overrides on value_type depending on
model choice selection.}
}
\value{
Epidemiolgical return values, either in cases or incidence, depending
 on user settings.
}
\description{
Calculate the epidemiological value to be shown as result, and on the reports.
}
