% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_validation.R
\name{calc_skill}
\alias{calc_skill}
\title{Calculate the forecast model skill score compared to the naive model predictions.}
\usage{
calc_skill(val_list, grp = NULL)
}
\arguments{
\item{val_list}{A list of 3 datasets of validation results: the first is the forecast model, the following two are the naive model results, as created by binding the results of calc_val_stats() in run_validation().}

\item{grp}{Optional inclusion of quo_groupfield when calculating skill scores by groupfield.}
}
\value{
Single dataset with skill scores of the main forecast model against each of the naive models, per timestep ahead, and optionally, per geographic grouping
}
\description{
Calculate the forecast model skill score compared to the naive model predictions.
}
