% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_validation.R
\name{calc_skill_stat}
\alias{calc_skill_stat}
\title{Calculate model skill comparison statistics}
\usage{
calc_skill_stat(fc_stat, naive_stat, perfect_stat)
}
\arguments{
\item{fc_stat}{The forecast model statistic value.}

\item{naive_stat}{The naive model statistic value (same statistic as forecast model).}

\item{perfect_stat}{The value of a perfect score for that stastistic.}
}
\value{
Skill score: the relative improvement the forecast model has over the naive model.
}
\description{
Helper function to calculate the relative improvement of the forecast over the specified naive model.
Skill score = (score_fc - score_naive) / (score_perfect - score_naive)
Skill metric has an upper bound of 1. No improvement is 0. Lower bound depends on statistic.
}
