% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecasting_helpers.R
\name{check_bb_models}
\alias{check_bb_models}
\title{Checks that all models successfully built when using batch_bam.}
\usage{
check_bb_models(reg_bb)
}
\arguments{
\item{reg_bb}{The regression model(s) returned by batch_bam.}
}
\value{
None. Will stop will informative error message if
problems are found.
}
\description{
Checks that all models successfully built when using batch_bam.
}
