% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatters_calculators.R
\name{create_epi_summary}
\alias{create_epi_summary}
\title{Creates summary of incidence in ED period}
\usage{
create_epi_summary(obs_res, quo_groupfield, report_dates)
}
\description{
Creates summary of incidence in ED period
}
