% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecasting.R
\name{create_predictions}
\alias{create_predictions}
\title{Create the appropriate predictions/forecasts.}
\usage{
create_predictions(model_choice, regress, epi_lag, req_date, ncores)
}
\arguments{
\item{model_choice}{Critical argument to choose the type of model to generate.
The options are versions that the EPIDEMIA team has used for forecasting.
The first supported options is "poisson-gam" ("p") which is the original
epidemiar model: a Poisson regression using bam (for large data GAMs), with
a smoothed cyclical for seasonality. The default for fc_control$anom_env is
TRUE for using the anomalies of environmental variables rather than their
raw values. The second option is "negbin" ("n") which is a negative binomial
regression using glm, with no external seasonality terms - letting the
natural cyclical behavior of the environmental variables fill that role. The
default for fc_control$anom_env is FALSE and uses the actual observation
values in the modeling. The fc_control$anom_env can be overruled by the user
providing a value, but this is not recommended unless you are doing
comparisons.}

\item{regress}{The regression object, either the user-supplied one, or
the one just generated.}

\item{epi_lag}{Epidemiological dataset with basis spline summaries of the
lagged environmental data (or anomalies), with groupings as a factor.}

\item{req_date}{The end date of requested forecast regression. When fit_freq
== "once", this is the last date of the full report, the end date of the
forecast period.}

\item{ncores}{The number of physical cores to use in parallel processing, set
in fc_control$ncores, else the max of the number of physical core available
minus 1, or 1 core.}
}
\value{
A dataset from predict() using the regression object generated in
 build_model or a newly created one. The dataset includes the
 predicted/forecast values through the end of the report requested.
}
\description{
Create the appropriate predictions/forecasts.
}
