% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_to_daily.R
\name{data_to_daily}
\alias{data_to_daily}
\title{With any interval or missing data, converts into daily values
with optional linear interpolation}
\usage{
data_to_daily(data_notdaily, valuefield, interpolate = TRUE)
}
\description{
With any interval or missing data, converts into daily values
with optional linear interpolation
}
