% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecasting.R
\name{env_last_week_mean}
\alias{env_last_week_mean}
\title{Calculate mean of last week environmental values.}
\usage{
env_last_week_mean(env_df, env_variables_used, quo_groupfield,
  quo_obsfield, groupings)
}
\arguments{
\item{env_df}{An environmental dataset with NA values at the end of known data.}

\item{env_variables_used}{List of environmental variables that were used in
the modeling.}

\item{quo_groupfield}{Quosure of the user given geographic grouping field to
run_epidemia().}

\item{quo_obsfield}{Quosure of user given field name of the environmental data
variables.}

\item{groupings}{A unique list of the geographic groupings (from groupfield).}
}
\value{
Data table. The first NA slot of an environmental variable per
 grouping is filled in with the mean of the previous week of daily
 environmental data.
}
\description{
Calculate mean of last week environmental values.
}
