% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatters_calculators.R
\name{environ_report_format}
\alias{environ_report_format}
\title{Formats env data for report}
\usage{
environ_report_format(env_ext_data, env_ref_data, quo_groupfield,
  quo_obsfield, env_used, env_info, week_type, report_dates)
}
\description{
Formats env data for report
}
