% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleaners_helpers.R
\name{epi_NA_interpolate}
\alias{epi_NA_interpolate}
\title{Interpolates missing epi data}
\usage{
epi_NA_interpolate(epi_data, quo_casefield, quo_groupfield)
}
\description{
Interpolates missing epi data
}
