% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epiwday.R
\name{epiwday}
\alias{epiwday}
\title{Get epidemiological weekday number of a date-time}
\usage{
epiwday(x, system = "ISO")
}
\arguments{
\item{x}{a date-time object. Must be a POSIXct, POSIXlt, Date, chron,
yearmon, yearqtr, zoo, zooreg, timeDate, xts, its, ti, jul, timeSeries, or
fts object.}

\item{system}{String indicating the standard (WHO ISO-8601 or CDC epi
weeks) that the weeks of the year in epidemiological and environmental
reference data use ["ISO" or "CDC"]. (Required: epidemiological observation
dates listed are LAST day of week).#'}
}
\value{
The weekday number (1--7) as an integer vector.
}
\description{
Returns the epidemiological weekday number using either ISO or CDC
system.
}
\details{
The WHO system uses the ISO 8601 standard in which weeks start on Monday,
while in the CDC system weeks start on Sunday.
}
\examples{
epiwday(as.Date("2005-01-01")) # 6
epiwday(as.Date("2005-01-01"), system = "ISO") # 6
epiwday(as.Date("2005-01-01"), system = "CDC") # 7

}
\references{
\url{https://en.wikipedia.org/wiki/ISO_week_date}
\url{https://www.cmmcp.org/sites/cmmcp/files/uploads/spring_skeeter_06.pdf}
}
