% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epiweek.R
\name{epiweek}
\alias{epiweek}
\title{Get epidemiological week number of a date-time}
\usage{
epiweek(x, system = "who")
}
\arguments{
\item{x}{a date-time object. Must be a POSIXct, POSIXlt, Date, chron,
yearmon, yearqtr, zoo, zooreg, timeDate, xts, its, ti, jul, timeSeries, or
fts object.}

\item{system}{Either "who" or "cdc". WHO epidemiological weeks start on
Monday. CDC epidemiological weeks (MMWR weeks) start on Sunday. The default
is "who".}
}
\value{
the weeks element of x as an integer number
}
\description{
Returns the epidemiological week number using either the WHO or CDC system,
both of which use a reoccuring leap year resulting in week numbers from
1--53.
}
\details{
The WHO system uses the ISO 8601 standard in which weeks start on Monday,
while in the CDC system weeks start on Sunday.

In both systems, week 1 of a given year is identified as being the first week
of the year with at least four days in that calendar year. As a result, the
epidemiological year may differ from the Gregorian calendar year for some
dates near January 1.

Internally, this function uses [lubridate::isoweek()] for WHO weeks and a
modified version of that functions code for CDC weeks.
}
\examples{
epiweek(as.Date("2005-01-01")) # 52
epiweek(as.Date("2005-01-01"), system = "cdc") # 1

}
\references{
\url{http://en.wikipedia.org/wiki/ISO_week_date}
\url{http://www.cmmcp.org/epiweek.htm}
}
