% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecasting.R
\name{extend_epi_future}
\alias{extend_epi_future}
\title{Extend epidemiology dataframe into future}
\usage{
extend_epi_future(epi_data, quo_popfield, quo_groupfield, groupings,
  report_dates)
}
\description{
Extend epidemiology dataframe into future
}
