% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecasting_helpers.R
\name{format_lag_ca}
\alias{format_lag_ca}
\title{Formats the environmental data lagged to epidemiology data the way that the
clusterapply package wants.}
\usage{
format_lag_ca(tbl, env_variables_used)
}
\arguments{
\item{tbl}{The tibble with all the lagged variables wide and flat.}

\item{env_variables_used}{Vector of the names of the environmental variables
that are being used in the model.}
}
\value{
A dataframe with sub-matrices for each of the lagged environmental
 variable data.
}
\description{
Formats the environmental data lagged to epidemiology data the way that the
clusterapply package wants.
}
