% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatters_calculators.R
\name{format_report_settings}
\alias{format_report_settings}
\title{Formats report_settings for including in metadata part of final report data}
\usage{
format_report_settings(rpt_settings)
}
\arguments{
\item{rpt_settings}{Report settings after processing defaults and matching.}
}
\value{
Named list of report_settings, in alphabetical order and no developer settings
}
\description{
Formats report_settings for including in metadata part of final report data
}
