% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_validation.R
\name{get_group_validations}
\alias{get_group_validations}
\title{Get geographic grouping model validation statistics}
\usage{
get_group_validations(validations)
}
\arguments{
\item{validations}{The set of validation statistics produced by
run_validation() - only the list of validation data sets, not including the skill metrics.}
}
\value{
A list of tibbles containing only the model geographic grouping statistics.
}
\description{
Small function to pull out just the geographic grouping validation statistics.
}
