% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_validation.R
\name{get_overall_validations}
\alias{get_overall_validations}
\title{Get overall model validation statistics}
\usage{
get_overall_validations(validations)
}
\arguments{
\item{validations}{The set of validation statistics produced by
run_validation() - only the list of validation data sets, not including the skill metrics.}
}
\value{
A list of tibbles containing only the model overall statistics (and
  not including the geographic grouping results, if present).
}
\description{
Small function to pull out just overall validation statistics.
}
