% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_checks.R
\name{input_check}
\alias{input_check}
\title{Functions to check input to epidemiar}
\usage{
input_check(epi_data, quo_casefield, quo_popfield, inc_per, quo_groupfield,
  week_type, report_period, ed_summary_period, ed_method, ed_control,
  env_data, quo_obsfield, quo_valuefield, forecast_future, fc_control,
  env_ref_data, env_info)
}
\arguments{
\item{epi_data}{Epidemiological data with case numbers per week, with date
field "obs_date".}

\item{quo_popfield}{Quosure of user-given field containing population values.}

\item{inc_per}{Number for what unit of population the incidence should be
reported in, e.g. incidence rate of 3 per 1000 people.}

\item{quo_groupfield}{Quosure of the user given geographic grouping field to
run_epidemia().}

\item{week_type}{String indicating the standard (WHO ISO-8601 or CDC epi
weeks) that the weeks of the year in epidemiological and environmental
reference data use ["ISO" or "CDC"]. (Required: epidemiological observation
dates listed are LAST day of week).}

\item{report_period}{The number of weeks that the entire report will cover.
The \code{report_period} minus \code{forecast_future} is the number of weeks
of past (known) data that will be included.}

\item{ed_summary_period}{The number of weeks that will be considered the "early
detection period". It will count back from the week of last known
epidemiological data.}

\item{ed_method}{Which method for early detection should be used ("Farrington"
is only current option, or "None").}

\item{ed_control}{All parameters for early detection algorithm, passed through
to that subroutine.}

\item{env_data}{Daily environmental data for the same groupfields and date
range as the epidemiological data. It may contain extra data (other
districts or date ranges). The data must be in long format (one row for each
date and environmental variable combination), and must start at absolutel
minimum \code{laglen} (in \code{fc_control}) days before epi_data for
forecasting.}

\item{quo_obsfield}{Quosure of user given field name of the environmental data
variables}

\item{quo_valuefield}{Quosure of user given field name of the value of the
environmental data variable observations.}

\item{forecast_future}{Number of futre weeks from the end of the
\code{epi_data} to produce forecasts.}

\item{fc_control}{Parameters for forecasting, including which environmental
variable to include and any geographic clusters.}

\item{env_ref_data}{Historical averages by week of year for environmental
variables. Used in extended environmental data into the future for long
forecast time, to calculate anomalies in early detection period, and to
display on timeseries in reports.}

\item{env_info}{Lookup table for environmental data - reference creation
method (e.g. sum or mean), report labels, etc.}
}
\value{
Returns a flag if there were any errors, plus accompanying error messages.
Also returns a flag and messages for warnings, as well.
}
\description{
Function does basic existance checks and variety of logic checks on input data to run_epidemia().
}
