% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecasting.R
\name{lag_environ_to_epi}
\alias{lag_environ_to_epi}
\title{Lag the env data}
\usage{
lag_environ_to_epi(epi_fc, quo_groupfield, groupings, env_fc,
  env_variables_used, laglen)
}
\description{
Lag the env data
}
