% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/event_detection.R
\name{make_stss}
\alias{make_stss}
\title{Make the list of sts objects}
\usage{
make_stss(epi_fc_data, quo_groupfield, quo_popfield, groupings)
}
\arguments{
\item{epi_fc_data}{Internal pass of epidemiological data complete with future
forecast values.}

\item{quo_groupfield}{Quosure of the user given geographic grouping field to
run_epidemia().}

\item{quo_popfield}{Quosure of user-given field containing population values.}

\item{groupings}{A unique list of the geographic groupings (from groupfield).}
}
\value{
A list of surveillance time series (sts) objects,
one for each geographic grouping.
}
\description{
Make the list of sts objects
}
