% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleaners_helpers.R
\name{Mode}
\alias{Mode}
\title{Mode}
\usage{
Mode(x, na.rm = FALSE)
}
\arguments{
\item{x}{A vector.}

\item{na.rm}{Logical indicating whether \code{NA} values should be excluded.}
}
\value{
A vector of length 1 and the same class as \code{x}.
}
\description{
Calculate the mode of a set of values, for numeric or character/factor data.
In ties, returns the first tied value.
}
\examples{
Mode(c(1,1,2,3))
Mode(c(1,2,2,3))
Mode(c(1,1,3,3))
Mode(c(3,3,1,1))
Mode(c(1,NA,NA))
Mode(c(1,NA,NA), na.rm = TRUE)

}
