% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/na_approx.R
\name{na_approx}
\alias{na_approx}
\title{Replace NA by interpolation in a numeric vector}
\usage{
na_approx(x, order.by = zoo::index(x))
}
\arguments{
\item{x}{A numeric vector.}

\item{order.by}{An index vector with unique entries by which the observations
in `x`` are ordered.}
}
\value{
A numeric vector of the same length as `x`.
}
\description{
Fill in missing values in a numeric vector by approximat them through linear
interpolation.
}
\details{
Leading and trailing `NA`s are left as is.
}
\examples{
na_approx(c(NA, .31, 4, NA, NA, 10, NA))
}
