% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/early_detection.R
\name{run_early_detection}
\alias{run_early_detection}
\title{Run early detection algorithm}
\usage{
run_early_detection(epi_data, quo_popfield, quo_groupfield, groupings,
  ed_method, ed_control, report_dates)
}
\description{
Run early detection algorithm
}
