% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/event_detection.R
\name{run_event_detection}
\alias{run_event_detection}
\title{Run event detection algorithm.
Generates three series:
"ed" : early detection alerts (ed period of most recent epi data)
"ew" : early warning alerts (forecast/future portion)
"thresh" : threshold values per week}
\usage{
run_event_detection(epi_fc_data, quo_popfield, inc_per, quo_groupfield,
  groupings, ed_method, ed_control, report_dates)
}
\description{
Run event detection algorithm.
Generates three series:
"ed" : early detection alerts (ed period of most recent epi data)
"ew" : early warning alerts (forecast/future portion)
"thresh" : threshold values per week
}
