% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/early_detection.R
\name{run_farrington}
\alias{run_farrington}
\title{Run the Farrington early detection algorithm}
\usage{
run_farrington(epi_fc_data, quo_popfield, quo_groupfield, groupings,
  ed_control, report_dates)
}
\description{
Run the Farrington early detection algorithm
}
