% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/event_detection.R
\name{run_no_detection}
\alias{run_no_detection}
\title{Run No outbreak detection algorithm}
\usage{
run_no_detection(epi_fc_data, quo_groupfield, report_dates)
}
\description{
Run No outbreak detection algorithm
}
