% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_validation.R
\name{save_geog_validations}
\alias{save_geog_validations}
\title{Save geographic grouping model validation statistics}
\usage{
save_geog_validations(validations, save_file)
}
\arguments{
\item{validations}{The set of validation statistics produced by
run_validation() - only the list of validation data sets, not including the skill metrics.}

\item{save_file}{File name to save results into csv format}
}
\value{
A csv file containing the model validation statistics for the
  geographic grouping results.
}
\description{
Small function to pull out validation statistics per geographic grouping and
save to csv.
}
