% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_validation.R
\name{save_overall_validations}
\alias{save_overall_validations}
\title{Save overall model validation statistics}
\usage{
save_overall_validations(validations, save_file)
}
\arguments{
\item{validations}{The set of validation statistics produced by
run_validation() - only the list of validation data sets, not including the skill metrics.}

\item{save_file}{File name to save results into csv format}
}
\value{
A csv file containing only the model overall statistics (and not
  including the geographic grouping results, if present).
}
\description{
Small function to pull out just overall validation statistics and save to
csv.
}
