% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_checks.R
\name{set_report_defaults}
\alias{set_report_defaults}
\title{Set defaults of any missing report_settings parameters}
\usage{
set_report_defaults(raw_settings, env_info, env_ref_data, env_variables,
  quo_obsfield, groupings, quo_groupfield)
}
\arguments{
\item{raw_settings}{The report_settings object as given by the user.}

\item{env_info}{Lookup table for environmental data - reference creation
method (e.g. sum or mean), report labels, etc.}

\item{env_ref_data}{Historical averages by week of year for environmental
variables. Used in extended environmental data into the future for long
forecast time, to calculate anomalies in early detection period, and to
display on timeseries in reports.}

\item{env_variables}{List of all unique environmental variables in env_data.}

\item{quo_obsfield}{Quosure of user given field name of the environmental data
variables.}

\item{groupings}{List of all unique geographical groupings in epi_data.}

\item{quo_groupfield}{Quosure of the user given geographic grouping field to
run_epidemia().}
}
\value{
Returns a full report_settings object, using user supplied values or
 defaults is option was missing.
}
\description{
Function sets defaults to report_settings parameters.
}
