% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/early_detection.R
\name{stss_res_to_output_data}
\alias{stss_res_to_output_data}
\title{Formats output data from sts result objects}
\usage{
stss_res_to_output_data(stss_res_list, quo_popfield, quo_groupfield, groupings,
  report_dates)
}
\description{
Formats output data from sts result objects
}
