% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecasting_helpers.R
\name{truncpoly}
\alias{truncpoly}
\title{Creates a modified b-spline basis (piecewise polynomial).}
\usage{
truncpoly(x = NULL, degree = 6, maxobs = NULL, minobs = NULL)
}
\arguments{
\item{x}{Vector of weekly observation dates.}

\item{degree}{Degree passed to splines::bs().}

\item{maxobs}{Date of the last known value.}

\item{minobs}{Date of the first known value.}
}
\value{
A modified b-spline basis with the last basis splines reversed and
 the second to last basis spline function removed (to reduce the edge effects
 of using splines).
}
\description{
The modified basis splines are used to capture any long term trends per
geographic group.
}
