% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_validation.R
\name{view_overall_validations}
\alias{view_overall_validations}
\title{View overall model validation statistics}
\usage{
view_overall_validations(validations)
}
\arguments{
\item{validations}{The set of validation statistics produced by
run_validation()}
}
\value{
A list of tibbles containing only the model overall statistics (and
  not including the geographic grouping results, if present).
}
\description{
Small function to pull out just overall validation statistics.
}
