% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_datefields.R
\name{add_datefields}
\alias{add_datefields}
\title{Adds ISO or CDC datefields to a data set with a "obs_date" (Date) field}
\usage{
add_datefields(df, type = "ISO")
}
\arguments{
\item{df}{Data table or tibble with a field named "obs_date"}

\item{type}{String to indicate whether to use "ISO" ISO-8601 week of year
(used by WHO) or "CDC" epi weeks.}
}
\value{
dataframe (or tibble) with week (`week_epidemiar`) and year
  (`year_epidemiar`) fields added
}
\description{
Adds ISO or CDC datefields to a data set with a "obs_date" (Date) field
}
