% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecasting.R
\name{anomalize_env}
\alias{anomalize_env}
\title{Convert environmental data into anomalies}
\usage{
anomalize_env(env_fc, quo_groupfield, env_variables_used, ncores)
}
\description{
Convert environmental data into anomalies
}
