% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecasting_main.R
\name{build_model}
\alias{build_model}
\title{Build the appropriate model}
\usage{
build_model(fc_model_family, quo_groupfield, epi_known, nthreads,
  n_groupings, modb_eq, bandsums_eq)
}
\arguments{
\item{fc_model_family}{model choice stand in <<>>}

\item{quo_groupfield}{Quosure of the user given geographic grouping field to
run_epidemia().}

\item{epi_known}{Epidemiological dataset with basis spline summaries of the
lagged environmental data (or anomalies), with column marking if "known"
data and groupings converted to factors.}

\item{nthreads}{thread count <<>>}

\item{n_groupings}{Count of the number of geographic groupings in the model.}

\item{modb_eq}{Pieces of the regression formula that include the modified
basis functions to account for long term trend (with or without groupings,
as appropriate).}

\item{bandsums_eq}{Pieces of the regression formula that include the b-spline
bandsummaries of the environmental factors.}
}
\value{
Regression object
}
\description{
Build the appropriate model
}
