% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatters_calculators.R
\name{calc_env_anomalies}
\alias{calc_env_anomalies}
\title{Calculates and summarizes environmental anomalies in the early detection period.}
\usage{
calc_env_anomalies(env_ts, quo_groupfield, quo_obsfield, report_dates)
}
\arguments{
\item{env_ts}{Environmental timeseries dataset as output from environ_report_format().}

\item{quo_groupfield}{Quosure of the user given geographic grouping field to
run_epidemia().}

\item{quo_obsfield}{Quosure of user given field name of the environmental data
variables.}

\item{report_dates}{Internally generated set of report date information: min,
max, list of dates for full report, known epidemiological data period,
forecast period, and early detection period.}
}
\value{
These data are the recent (during the early detection period)
  differences (anomalies) of the environmental variable values from the
  climatology/reference mean. Note: these are not the same as daily
  environmental anomalies calculated as residuals from GAM in anomalize_env()
  as part of forecasting.
}
\description{
Calculates and summarizes environmental anomalies in the early detection period.
}
