% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatters_calculators.R
\name{create_epi_summary}
\alias{create_epi_summary}
\title{Creates summary of disease incidence in early detection period.}
\usage{
create_epi_summary(obs_res, quo_groupfield, report_dates)
}
\arguments{
\item{obs_res}{Formatted dataset of observed disease incidence.}

\item{quo_groupfield}{Quosure of the user given geographic grouping field to
run_epidemia().}

\item{report_dates}{Internally generated set of report date information: min,
max, list of dates for full report, known epidemiological data period,
forecast period, and early detection period.}
}
\value{
Mean disease incidence per geographic group during the early detection
 period, returned as epi_summary in the run_epidemia() ouput.
}
\description{
Creates summary of disease incidence in early detection period.
}
