% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecasting_main.R
\name{create_predictions}
\alias{create_predictions}
\title{Create the appropriate predictions/forecasts.}
\usage{
create_predictions(fc_model_family, nthreads, regress, epi_lag, req_date)
}
\arguments{
\item{fc_model_family}{model choice <<>>}

\item{nthreads}{max threads <<>>}

\item{regress}{The regression object, either the user-supplied one, or
the one just generated.}

\item{epi_lag}{Epidemiological dataset with basis spline summaries of the
lagged environmental data (or anomalies), with groupings as a factor.}

\item{req_date}{The end date of requested forecast regression. When fit_freq
== "once", this is the last date of the full report, the end date of the
forecast period.}
}
\value{
A dataset from predict() using the regression object generated in
 build_model or a newly created one. The dataset includes the
 predicted/forecast values through the end of the report requested.
}
\description{
Create the appropriate predictions/forecasts.
}
