% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecasting_main.R
\name{create_predictions}
\alias{create_predictions}
\title{Create the appropriate predictions/forecasts.}
\usage{
create_predictions(fc_model_family, nthreads, regress, epi_lag, req_date)
}
\arguments{
\item{fc_model_family}{The \code{\link[stats]{family}} parameter passsed to
\code{\link[mgcv:bam]{mgcv::bam}}, and the extended families in
\code{\link[mgcv]{family.mgcv}} can also be used. This sets the type of
generalized additive model (GAM) to run: it specifies the distribution and
link to use in model fitting. E.g. for a Poisson regression, the user would
input "poisson()". If a cached model is being used, set the parameter to `"cached"`.}

\item{nthreads}{Extract of `report_settings$fc_nthreads`}

\item{regress}{The regression object, either the user-supplied one from
`report_settings$model_cached`, or the one just generated.}

\item{epi_lag}{Epidemiological dataset with basis spline summaries of the
lagged environmental data (or anomalies), with groupings as a factor.}

\item{req_date}{The end date of requested forecast regression. When fit_freq
== "once", this is the last date of the full report, the end date of the
forecast period.}
}
\value{
A dataset from predict() using the regression object generated in
 build_model or a newly created one. The dataset includes the
 predicted/forecast values through the end of the report requested.
}
\description{
Create the appropriate predictions/forecasts.
}
