% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecasting.R
\name{env_last_week_mean}
\alias{env_last_week_mean}
\title{Calculate mean of last week env values}
\usage{
env_last_week_mean(env_df, env_variables_used, quo_groupfield, quo_obsfield,
  groupings)
}
\description{
Calculate mean of last week env values
}
