% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatters_calculators.R
\name{environ_report_format}
\alias{environ_report_format}
\title{Formats environmental data for report timeseries.}
\usage{
environ_report_format(
  env_ext_data,
  env_ref_data,
  quo_groupfield,
  quo_obsfield,
  env_used,
  env_info,
  epi_date_type,
  report_dates
)
}
\arguments{
\item{env_ext_data}{An environmental dataset extended into the
future/forecast period with estimated values for the environmental
variables. The env_data_extd object returned by run_forecast().}

\item{env_ref_data}{Historical averages by week of year for environmental
variables. Used in extended environmental data into the future for long
forecast time, to calculate anomalies in early detection period, and to
display on timeseries in reports.}

\item{quo_groupfield}{Quosure of the user given geographic grouping field to
run_epidemia().}

\item{quo_obsfield}{Quosure of user given field name of the environmental data
variables}

\item{env_used}{List of environmental variables that were used in
the modeling.}

\item{env_info}{Lookup table for environmental data - reference creation
method (e.g. sum or mean), report labels, etc.}

\item{epi_date_type}{String indicating the standard (WHO ISO-8601 or CDC epi
weeks) that the weeks of the year in epidemiological and environmental
reference data use ["ISO" or "CDC"]. <<>>}

\item{report_dates}{Internally generated set of report date information: min,
max, list of dates for full report, known epidemiological data period,
forecast period, and early detection period.}
}
\value{
Data set of multiple timeseries for the used environmental variables
 during the report period for each geographic unit. Returned as
 environ_timeseries in the run_epidemia() output.
}
\description{
Formats environmental data for report timeseries.
}
