% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecasting_helpers.R
\name{epi_format_fc}
\alias{epi_format_fc}
\title{Format epi data for modeling}
\usage{
epi_format_fc(epi_data_extd, quo_groupfield, fc_clusters)
}
\arguments{
\item{epi_data_extd}{An epidemiological dataset extended into the
future/forecast period with NA values for to-be-forecasted case numbers.}

\item{quo_groupfield}{Quosure of the user given geographic grouping field to
run_epidemia().}

\item{fc_clusters}{Extract from `report_settings$fc_clusters`, the
geographic clusters to use in modeling.}
}
\value{
An epidemiological dataset formatted to pass over to BAM/GAM modeling.
}
\description{
Format epi data for modeling
}
