% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleaners_helpers.R
\name{epi_NA_interpolate}
\alias{epi_NA_interpolate}
\title{Interpolates missing epi data.}
\usage{
epi_NA_interpolate(epi_data, quo_casefield, quo_groupfield)
}
\arguments{
\item{epi_data}{Epidemiological data with case numbers per week, with date
field "obs_date".}

\item{quo_casefield}{Quosure of user given casefield to run_epidemia().}

\item{quo_groupfield}{Quosure of the user given geographic grouping field to
run_epidemia().}
}
\value{
Same data as epi_data with new interpolated case field,
  cases_epidemiar.
}
\description{
Interpolates missing epi data.
}
