% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecasting.R
\name{extend_env_future}
\alias{extend_env_future}
\title{Extend environmental data into the future}
\usage{
extend_env_future(env_data, quo_groupfield, groupings, quo_obsfield,
  quo_valuefield, env_ref_data, env_info, env_variables_used, report_dates,
  week_type)
}
\description{
Extend environmental data into the future
}
