% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecasting_helpers.R
\name{extend_epi_future}
\alias{extend_epi_future}
\title{Extend epidemiology dataframe into future.}
\usage{
extend_epi_future(
  epi_data,
  quo_popfield,
  quo_groupfield,
  groupings,
  report_dates
)
}
\arguments{
\item{epi_data}{Epidemiological data with case numbers per week, with date
field "obs_date".}

\item{quo_popfield}{Quosure of user-given field containing population values.}

\item{quo_groupfield}{Quosure of the user given geographic grouping field to
run_epidemia().}

\item{groupings}{A unique list of the geographic groupings (from groupfield).}

\item{report_dates}{Internally generated set of report date information: min,
max, list of dates for full report, known epidemiological data period,
forecast period, and early detection period.}
}
\value{
Epidemiological dataset extended past the known epi data time range
 and into the future/forecast period. Case numbers are filled in the NA (to
 be forecasted), and the population is estimated in a persistence method.
}
\description{
Extend epidemiology dataframe into future.
}
