% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecasting_main.R
\name{forecast_regression}
\alias{forecast_regression}
\title{Run forecast regression}
\usage{
forecast_regression(epi_lag, quo_groupfield, fc_model_family,
  report_settings, groupings, env_variables_used, report_dates, req_date)
}
\arguments{
\item{epi_lag}{Epidemiological dataset with basis spline summaries of the
lagged environmental data (or anomalies), as output by lag_environ_to_epi().}

\item{quo_groupfield}{Quosure of the user given geographic grouping field to
run_epidemia().}

\item{fc_model_family}{model choice stand in}

\item{report_settings}{report settings}

\item{groupings}{A unique list of the geographic groupings (from groupfield).}

\item{env_variables_used}{List of environmental variables that were used in
the modeling.}

\item{report_dates}{Internally generated set of report date information: min,
max, list of dates for full report, known epidemiological data period,
forecast period, and early detection period.}

\item{req_date}{The end date of requested forecast regression. When fit_freq
== "once", this is the last date of the full report, the end date of the
forecast period.}
}
\value{
Named list containing:
date_preds: Full forecasted resulting dataset.
reg_obj: The regression object from modeling.
Unless model_run is TRUE, in which case only the regression object is returned.
}
\description{
Run forecast regression
}
