% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecasting.R
\name{forecast_regression}
\alias{forecast_regression}
\title{Run forecast regression}
\usage{
forecast_regression(epi_lag, quo_groupfield, groupings, env_variables_used,
  report_dates, req_date, ncores, fit_freq, model_run, model_obj = NULL,
  model_cached = NULL, model_choice)
}
\arguments{
\item{epi_lag}{Epidemiological dataset with basis spline summaries of the
lagged environmental data anomalies, as output by lag_environ_to_epi().}

\item{quo_groupfield}{Quosure of the user given geographic grouping field to
run_epidemia().}

\item{groupings}{A unique list of the geographic groupings (from groupfield).}

\item{env_variables_used}{List of environmental variables that were used in
the modeling.}

\item{report_dates}{Internally generated set of report date information: min,
max, list of dates for full report, known epidemiological data period,
forecast period, and early detection period.}

\item{req_date}{The end date of requested forecast regression. When fit_freq
== "once", this is the last date of the full report, the end date of the
forecast period.}

\item{ncores}{The number of physical cores to use in parallel processing, set
in fc_control$ncores, else the max of the number of physical core available
minus 1, or 1 core.}

\item{fit_freq}{String indicating "once" or "weekly" on how often to fit the
model - once for the whole report, or every week of the report. Unless
otherwise needed, the value should be "once", as weekly drastically
increases processing time.}

\item{model_run}{TRUE/FALSE flag for whether to only generate the model
regression object plus metadata. This model can be cached and used later on
its own, skipping a large portion of the slow calculations for future runs.}

\item{model_obj}{Deprecated, use model_cached if possible. Regression object
built from a model_run = TRUE run of run_epidemia(). Using the prebuilt
model will significantly save on processing time, but will need to be
updated periodically.}

\item{model_cached}{The output of a previous model_run = TRUE run of
run_epidemia() that produces a model (regression object) and metadata. The
metadata will be used for input checking and validation. Using a prebuilt
model saves on processing time, but will need to be updated periodically.}

\item{model_choice}{Critical argument to choose the type of model to generate.
The options are versions that the EPIDEMIA team has used for forecasting.
The first supported options is "poisson-gam" ("p") which is the original
epidemiar model: a Poisson regression using bam (for large data GAMs), with
a smoothed cyclical for seasonality. The default for fc_control$anom_env is
TRUE for using the anomalies of environmental variables rather than their
raw values. The second option is "negbin" ("n") which is a negative binomial
regression using glm, with no external seasonality terms - letting the
natural cyclical behavior of the environmental variables fill that role. The
default for fc_control$anom_env is FALSE and uses the actual observation
values in the modeling. The fc_control$anom_env can be overruled by the user
providing a value, but this is not recommended unless you are doing
comparisons.}
}
\value{
Named list containing:
date_preds: Full forecasted resulting dataset.
reg_obj: The regression object from modeling.
Unless model_run is TRUE, in which case only the regression object is returned.
}
\description{
Run forecast regression
}
