% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecasting.R
\name{forecast_regression}
\alias{forecast_regression}
\title{Run forecast regression}
\usage{
forecast_regression(epi_lag, quo_groupfield, groupings, env_variables_used,
  req_date, ncores, fit_freq, rpt_start)
}
\description{
Run forecast regression
}
