% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_date_yw.R
\name{make_date_yw}
\alias{make_date_yw}
\title{Create dates from epidemiological weeks and years}
\usage{
make_date_yw(year = 1970L, week = 1L, weekday = 1L, system = "ISO")
}
\arguments{
\item{year}{epidemiological year}

\item{week}{eidemiological week number (1--53).}

\item{weekday}{epidemiological weekday number (1--7). Day 1 is a Monday in
the ISO-8601 WHO system and a Sunday in the CDC system.}

\item{system}{String indicating the standard (WHO ISO-8601 or CDC epi
weeks) that the weeks of the year in epidemiological and environmental
reference data use ["ISO" or "CDC"]. (Required: epidemiological observation
dates listed are LAST day of week).#'}
}
\value{
A vector of class `Date`.
}
\description{
Calculate Gregorian calendar dates based on epidemiological years, week
numbers, and weekday numbers.
}
\details{
Arguments `year`, `week`, and `weekday` are recycled as necessary.
}
\examples{
make_date_yw(2017, 1)
make_date_yw(2017, 1, 2)
make_date_yw(2017, 1, system = "CDC")
make_date_yw(2017, 1, system = "ISO")
make_date_yw(2017, 1, 2, system = "ISO")

# arguments are recycled
make_date_yw(2017, 1:10)
make_date_yw(2017, 1, 1:7)
make_date_yw(2010:2017, 1)

}
\references{
\url{https://en.wikipedia.org/wiki/ISO_week_date}
\url{https://www.cmmcp.org/sites/cmmcp/files/uploads/spring_skeeter_06.pdf}
}
