% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecasting_helpers.R
\name{pull_model_envvars}
\alias{pull_model_envvars}
\title{Pull only model env variables.}
\usage{
pull_model_envvars(env_data, quo_obsfield, env_var)
}
\arguments{
\item{env_data}{Daily environmental data for the same groupfields and date
range as the epidemiological data. It may contain extra data (other
districts or date ranges). The data must be in long format (one row for each
date and environmental variable combination), and must start at absolutel
minimum \code{report_settings$env_lag_length} days (default 180) before
epi_data for forecasting.}

\item{quo_obsfield}{Quosure of user given field name of the environmental data
variables.}

\item{env_var}{Extract from report_settings$env_var, a user list of environmental variables to attempt to use for modeling}
}
\value{
List of environmental variables that were used in the
 modeling (had to be both listed in model variables input file and present the
 env_data dataset).
}
\description{
Pull only model env variables.
}
