% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_epidemia.R
\name{run_epidemia}
\alias{run_epidemia}
\title{Run EPIDEMIA forecast models and early detection algorithm}
\usage{
run_epidemia(epi_data, casefield, populationfield, inc_per = 1000,
  groupfield, week_type = c("ISO", "CDC"), report_period = 26,
  ed_summary_period = 4, ed_method = c("Farrington", "EARS"),
  ed_control = NULL, env_data, obsfield, valuefield,
  forecast_future = 4, fc_control = NULL, env_ref_data, env_info)
}
\arguments{
\item{epi_data}{Epidemiological data with case numbers per week, with date field "obs_date"}

\item{casefield}{The field name for the case counts}

\item{populationfield}{Population field to give population numbers over time. Used to calculated incidence. Also optionally used in Farrington method for populationOffset.}

\item{groupfield}{The field name for districts or area divisions of epidemiological AND environmental data if no groupings (all one area), user should give a field with the same value throughout}

\item{week_type}{The standard (WHO ISO-8601 or CDC epi weeks) that the weeks of the year in epidemiological and environmental reference data use (required: dates listed are LAST day of week).}

\item{report_period}{The number of timesteps (weeks) that the whole report will cover. report_period - forecast_future is the number of weeks of past (known) data that will be included.}

\item{ed_method}{Which method for early detection should be used ("Farrington" is only current) option.}

\item{ed_control}{All parameters for early detection algorithm, passed through to that subroutine}

\item{env_data}{Daily environmental data for same groupfields and Date range. Must be in long format, and must start laglen (in fc_control) days before epi_data for forecasting.}

\item{obsfield}{Field name of the environmental data observation types}

\item{valuefield}{Field name of the value of the environmental data observations}

\item{forecast_future}{Number of weeks from the end of the epi_data to extend the forecast out}

\item{fc_control}{Parameters for forecasting, including model and clusters}

\item{env_ref_data}{Historical averages by week of year for environmental variables. Used in extended environmental data into the future for long forecast time, to calculate anomalies in early detection period, and to display on timeseries in reports}

\item{env_info}{Lookup table for environmental data - reference creation method, report label, GA ID, etc.}

\item{detection_period}{The number of timesteps (weeks) that the early detection will run over}
}
\value{
Returns a suite of summary and report data.
#'
}
\description{
Run EPIDEMIA forecast models and early detection algorithm
}
\examples{

}
