% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_epidemia.R
\name{run_epidemia}
\alias{run_epidemia}
\title{Run EPIDEMIA forecast models and early detection algorithm.}
\usage{
run_epidemia(epi_data = NULL, env_data = NULL, env_ref_data = NULL,
  env_info = NULL, casefield = NULL, groupfield = NULL,
  populationfield = NULL, obsfield = NULL, valuefield = NULL,
  fc_model_family = NULL, report_settings = NULL)
}
\arguments{
\item{epi_data}{Epidemiological data with case numbers per week, with date
field "obs_date".}

\item{env_data}{Daily environmental data for the same groupfields and date
range as the epidemiological data. It may contain extra data (other
districts or date ranges). The data must be in long format (one row for each
date and environmental variable combination), and must start at absolutel
minimum \code{laglen} (in \code{fc_control}) days before epi_data for
forecasting.}

\item{env_ref_data}{Historical averages by week of year for environmental
variables. Used in extended environmental data into the future for long
forecast time, to calculate anomalies in early detection period, and to
display on timeseries in reports.}

\item{env_info}{Lookup table for environmental data - reference creation
method (e.g. sum or mean), report labels, etc.}

\item{casefield}{The column name of the field that contains disease case
counts (unquoted field name).}

\item{groupfield}{The column name of the field for district or geographic area
unit division names of epidemiological AND environmental data (unquoted
field name). If there are no groupings (all one area), user should give a
field that contains the same value throughout.}

\item{populationfield}{Column name of the population field to give population
numbers over time (unquoted field name). Used to calculated incidence. Also
optionally used in Farrington method for populationOffset.}

\item{obsfield}{Field name of the environmental data variables (unquoted field
name).}

\item{valuefield}{Field name of the value of the environmental data variable
observations (unquoted field name).}

\item{fc_model_family}{Critical argument to choose the type of model to generate.
The options are versions that the EPIDEMIA team has used for forecasting.
The first supported options is "poisson-bam" ("p") which is the original
epidemiar model: a Poisson regression using bam (for large data GAMs), with
a smoothed cyclical for seasonality. The default for fc_control$anom_env is
TRUE for using the anomalies of environmental variables rather than their
raw values. The second option is "negbin" ("n") which is a negative binomial
regression using glm, with no external seasonality terms - letting the
natural cyclical behavior of the environmental variables fill that role. The
default for fc_control$anom_env is FALSE and uses the actual observation
values in the modeling. The fc_control$anom_env can be overruled by the user
providing a value, but this is not recommended unless you are doing
comparisons.}

\item{report_settings}{Optional report settings.  <<>>}
}
\value{
Returns a suite of summary and report data.

 1. \code{summary_data}: Early detection and early warning alerts levels for
 each woreda. Early detection alerts (ed_alert_count) are alerts that are
 triggered during the early detection period, which is defined as the 4 most
 recent weeks of known epidemiology data. Similarly, early warning alerts
 (ew_alert_count) are alerts in the future forecast estimates. “High” level
 indicates two or more weeks in this period had incidences greater than the
 alert threshold, “Medium” means that one week was in alert status, and “Low”
 means no weeks had alerts (ed_sum_level and ew_level, respectively).

 2. \code{epi_summary}: Mean disease incidence per geographic group during
 the early detection period.

 3. \code{modeling_results_data}:These are multiple timeseries values for
 observed, forecast, and alert thresholds of disease incidence, over the
 report period, for each geographic unit. These data can be used in creating
 the individual geographic unit control charts.

 4. \code{environ_timeseries}: These are multiple timeseries for the
 environmental variables during the report period for each geographic unit.

 5. \code{environ_anomalies}: These data are the recent (during the early
 detection period) differences (anomalies) of the environmental variable
 values from the climatology/reference mean.

 6. \code{params_meta}: This lists dates, settings, and parameters that
 \code{run_epidemiar()} was called with.

 7. \code{regression_object}: This is the regression object from the general
 additive model (GAM, parallelized with BAM) regression. This is only for
 statistical investigation of the model, and is usually not saved (very large
 object).

 For more details see the vignette on the output data:
 \code{vignette("output-report-data", package = "epidemiar")}

 However, if \code{model_run = TRUE}, the function returns a list of two
 objects. The first, \code{model_obj} is the regression object from whichever
 model is being run. There is also \code{model_info} which has details on the
 parameters used to create the model, similar to \code{params_meta} in a full
 run.
}
\description{
The Epidemic Prognosis Incorporating Disease and Environmental Monitoring for
Integrated Assessment (EPIDEMIA) Forecasting System is a set of tools coded in
free, open-access software, that integrate surveillance and environmental data
to model and create short-term forecasts for environmentally-mediated
diseases. This function, `epidemiar::run_epidemia()` is the central function
to model and forecast a wide range of environmentally-mediated diseases.
}
\details{
For more a longer description of the package, see the overview vignette:
\code{vignette("overview-epidemiar", package = "epidemiar")}

For more details see the vignette on input data and modeling parameters:
\code{vignette("data-modeling", package = "epidemiar")}
}
\examples{
"See model_forecast_script in epidemiar-demo for full example:
https://github.com/EcoGRAPH/epidemiar-demo"

}
