% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_epidemia.R
\name{run_epidemia}
\alias{run_epidemia}
\title{Run EPIDEMIA forecast models and early detection algorithm.}
\usage{
run_epidemia(epi_data, casefield, populationfield, inc_per = 1000,
  groupfield, week_type = c("ISO", "CDC"), report_period = 26,
  ed_summary_period = 4, ed_method = c("Farrington", "None"),
  ed_control = NULL, env_data, obsfield, valuefield,
  forecast_future = 4, fc_control = NULL, env_ref_data, env_info)
}
\arguments{
\item{epi_data}{Epidemiological data with case numbers per week, with date
field "obs_date".}

\item{casefield}{The column name of the field that contains disease case
counts (unquoted field name).}

\item{populationfield}{Column name of the population field to give population
numbers over time (unquoted field name). Used to calculated incidence. Also
optionally used in Farrington method for populationOffset.}

\item{groupfield}{The column name of the field for district or geographic area
unit division names of epidemiological AND environmental data (unquoted
field name). If there are no groupings (all one area), user should give a
field that contains the same value throughout.}

\item{week_type}{The standard (WHO ISO-8601 or CDC epi weeks) that the weeks
of the year in epidemiological and environmental reference data use
(required: dates listed are LAST day of week).}

\item{report_period}{The number of weeks that the entire report will cover.
The \code{report_period} minus \code{forecast_future} is the number of weeks
of past (known) data that will be included.}

\item{ed_method}{Which method for early detection should be used ("Farrington"
is only current option, or "None").}

\item{ed_control}{All parameters for early detection algorithm, passed through
to that subroutine.}

\item{env_data}{Daily environmental data for the same groupfields and date
range as the epidemiological data. It may contain extra data (other
districts or date ranges). The data must be in long format (one row for each
date and environmental variable combination), and must start \code{laglen}
(in \code{fc_control}) days before epi_data for forecasting.}

\item{obsfield}{Field name of the environmental data variables (unquoted field
name).}

\item{valuefield}{Field name of the value of the environmental data variable
observations (unquoted field name).}

\item{forecast_future}{Number of futre weeks from the end of the
\code{epi_data} to produce forecasts.}

\item{fc_control}{Parameters for forecasting, including which environmental
variable to include and any geographic clusters.}

\item{env_ref_data}{Historical averages by week of year for environmental
variables. Used in extended environmental data into the future for long
forecast time, to calculate anomalies in early detection period, and to
display on timeseries in reports.}

\item{env_info}{Lookup table for environmental data - reference creation
method (e.g. sum or mean), report labels, etc.}

\item{detection_period}{The number of weeks that will be considered the "early
detection period". It will count back from the week of last known
epidemiological data.}
}
\value{
Returns a suite of summary and report data.

 1. \code{summary_data}: Early detection and early warning alerts levels for
 each woreda. Early detection alerts (ed_alert_count) are alerts that are
 triggered during the early detection period, which is defined as the 4 most
 recent weeks of known epidemiology data. Similarly, early warning alerts
 (ew_alert_count) are alerts in the future forecast estimates. “High” level
 indicates two or more weeks in this period had incidences greater than the
 alert threshold, “Medium” means that one week was in alert status, and “Low”
 means no weeks had alerts (ed_sum_level and ew_level, respectively).

 2. \code{epi_summary}: Mean disease incidence per geographic group during
 the early detection period.

 3. \code{modeling_results_data}:These are multiple timeseries values for
 observed, forecast, and alert thresholds of disease incidence, over the
 report period, for each geographic unit. These data can be used in creating
 the individual geographic unit control charts.

 4. \code{environ_timeseries}: These are multiple timeseries for the
 environmental variables during the report period for each geographic unit.

 5. \code{environ_anomalies}: These data are the recent (during the early
 detection period) differences (anomalies) of the environmental variable
 values from the climatology/reference mean.

 6. \code{params_meta}: This lists dates, settings, and parameters that
 \code{run_epidemiar()} was called with.

 7. \code{regression_object}: This is the regression object from the general
 additive model (GAM, parallelized with BAM) regression. This is only for
 statistical investigation of the model, and is usually not saved (very large
 object).

For more details see the vignette on the output data:
\code{vignette("output-report-data", package = "epidemiar")}
}
\description{
The Epidemic Prognosis Incorporating Disease and Environmental Monitoring for
Integrated Assessment (EPIDEMIA) Forecasting System is a set of tools coded in
free, open-access software, that integrate surveillance and environmental data
to model and create short-term forecasts for environmentally-mediated
diseases. This function, `epidemiar::run_epidemia()` is the central function
to model and forecast a wide range of environmentally-mediated diseases.
}
\details{
For more a longer description of the package, see the overview vignette:
\code{vignette("overview-epidemiar", package = "epidemiar")}

For more details see the vignette on input data and modeling parameters:
\code{vignette("data-modeling", package = "epidemiar")}
}
\examples{
"See model_forecast_script in epidemiar-demo for full example: https://github.com/EcoGRAPH/epidemiar-demo"

}
