% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/event_detection.R
\name{run_event_detection}
\alias{run_event_detection}
\title{Main subfunction for running event detection algorithm.}
\usage{
run_event_detection(
  epi_fc_data,
  quo_groupfield,
  quo_popfield,
  ed_method,
  ed_control,
  val_type,
  inc_per,
  groupings,
  report_dates,
  valid_run
)
}
\arguments{
\item{epi_fc_data}{Internal pass of epidemiological data complete with future
forecast values.}

\item{quo_groupfield}{Quosure of the user given geographic grouping field to
run_epidemia().}

\item{quo_popfield}{Quosure of user-given field containing population values.}

\item{ed_method}{An extract of report_settings$ed_method after defaults have
been applied - which method for early detection should be used ("farrington"
or default "none", currently).}

\item{ed_control}{An extract of report_settings$ed_control - all parameters
for early detection algorithm, passed through to that subroutine.}

\item{val_type}{An extract of report_settings$report_value_type after defaults
applies - whether to return epidemiological report values in "incidence" or
"cases" (default).}

\item{inc_per}{An extract of report_settings$report_inc_per after defaults
applies - number for what unit of population the incidence should be
reported in, e.g. incidence rate per 1000 people. Ignored when
report_settings$report_value_type is 'cases'.}

\item{groupings}{A unique list of the geographic groupings (from groupfield).}

\item{report_dates}{Internally generated set of report date information: min,
max, list of dates for full report, known epidemiological data period,
forecast period, and early detection period.}

\item{valid_run}{Internal TRUE/FALSE for whether this is part of a validation
run.}
}
\value{
Returns a list of three generated series: "ed" : early detection
 alerts (ed period of most recent epi data) "ew" : early warning alerts
 (forecast/future portion) "thresh" : threshold values per week
}
\description{
Main subfunction for running event detection algorithm.
}
