% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/event_detection.R
\name{run_no_detection}
\alias{run_no_detection}
\title{Run No outbreak detection algorithm}
\usage{
run_no_detection(epi_fc_data, quo_groupfield, report_dates)
}
\arguments{
\item{epi_fc_data}{Internal pass of epidemiological data complete with future
forecast values.}

\item{quo_groupfield}{Quosure of the user given geographic grouping field to
run_epidemia().}

\item{report_dates}{Internally generated set of report date information: min,
max, list of dates for full report, known epidemiological data period,
forecast period, and early detection period.}
}
\value{
Returns a list of three generated series with all NAs:
"ed" : early detection alerts (ed period of most recent epi data)
"ew" : early warning alerts (forecast/future portion)
"thresh" : threshold values per week
}
\description{
Run No outbreak detection algorithm
}
