% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/event_detection.R
\name{stss_res_to_output_data}
\alias{stss_res_to_output_data}
\title{Formats output data from sts result objects}
\usage{
stss_res_to_output_data(stss_res_list, epi_fc_data, quo_popfield, inc_per,
  quo_groupfield, groupings, report_dates)
}
\description{
Formats output data from sts result objects
}
